// src/api/__generated__/wallet/model/platformShareMetricRules.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { ShareMetricRule } from './shareMetricRule';

/**
 * 单平台传播数据积分规则
 */
export interface PlatformShareMetricRules {
  play?: ShareMetricRule;
  like?: ShareMetricRule;
  retweet?: ShareMetricRule;
}