// src/api/__generated__/wallet/model/submitShareAttemptRequestPlatform.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 分享平台（大小写不敏感）
 */
export type SubmitShareAttemptRequestPlatform =
  (typeof SubmitShareAttemptRequestPlatform)[keyof typeof SubmitShareAttemptRequestPlatform];

export const SubmitShareAttemptRequestPlatform = {
  x: 'x',
  tiktok: 'tiktok',
  youtube: 'youtube',
} as const;